Move mkString logic to FormatWriter.
BestFirstSearch should only implement the search.
