+++ core/src/main/scala/org/scalafmt/internal/FormatOps.scala
@@ -7,10 +7,13 @@ import org.scalafmt.ScalaStyle
 import org.scalafmt.util.TreeOps
 
 import scala.annotation.tailrec
+import scala.collection.mutable
+
 import scala.meta.Tree
 import scala.meta.internal.ast.Case
 import scala.meta.internal.ast.Template
 import scala.meta.internal.ast.Term
+import scala.meta.internal.ast.Type
 import scala.meta.tokens.Token
 import scala.meta.tokens.Token._
 import scala.meta.internal.ast.Defn
@@ -20,14 +23,14 @@ import scala.meta.prettyprinters.Structure
 /**
   * Helper functions for generating splits/policies for a given tree.
   */
-class FormatOps(val style: ScalaStyle,
-                val tree: Tree,
-                val tokens: Array[FormatToken],
-                val ownersMap: Map[TokenHash, Tree],
-                val statementStarts: Map[TokenHash, Tree],
-                val matchingParentheses: Map[TokenHash, Token]) {
+class FormatOps(val tree: Tree, val style: ScalaStyle) {
   import TokenOps._
   import TreeOps._
+
+  val tokens: Array[FormatToken] = FormatToken.formatTokens(tree.tokens)
+  val ownersMap = getOwners(tree)
+  val statementStarts = getStatementStarts(tree)
+  val matchingParentheses = getMatchingParentheses(tree.tokens)
 
   @inline
   def owners(token: Token): Tree = ownersMap(hash(token))
@@ -265,6 +268,7 @@ class FormatOps(val style: ScalaStyle,
       args.last.tokens.last
     }).getOrElse(owner.tokens.last)
   }
+
   def functionExpire(function: Term.Function): Token = {
     (for {
       parent <- function.parent
@@ -273,5 +277,30 @@ class FormatOps(val style: ScalaStyle,
         case _ => None
       }
     } yield blockEnd).getOrElse(function.tokens.last)
+  }
+
+  def noOptimizationZones(tree: Tree): Set[Token] = {
+    val result = new mutable.SetBuilder[Token, Set[Token]](Set.empty[Token])
+    var inside = false
+    var expire = tree.tokens.head
+    tree.tokens.foreach {
+      case t
+        if !inside &&
+            ((t, ownersMap(hash(t))) match {
+              case (_: `(`, _: Term.Apply) =>
+                // TODO(olafur) https://github.com/scalameta/scalameta/issues/345
+                val x = true
+                x
+              // Type compounds can be inside defn.defs
+              case (_: `{`, _: Type.Compound) => true
+              case _ => false
+            }) =>
+        inside = true
+        expire = matchingParentheses(hash(t))
+      case x if x == expire => inside = false
+      case x if inside => result += x
+      case _ =>
+    }
+    result.result()
   }
 }
