+++ core/src/main/scala/org/scalafmt/internal/FormatWriter.scala
@@ -1,0 +1,61 @@
+package org.scalafmt.internal
+
+import scala.meta.tokens.Token
+import scala.meta.tokens.Token._
+
+/**
+  * Produces formatted output from sequence of splits.
+  */
+class FormatWriter(formatOps: FormatOps) {
+  import formatOps._
+
+  def mkString(splits: Vector[Split]): String = {
+    val sb = new StringBuilder()
+    var lastState =
+      State.start // used to calculate start of formatToken.right.
+    State.reconstructPath(tokens, splits, style) {
+      case (state, formatToken, whitespace) =>
+        formatToken.left match {
+          case c: Comment if c.code.startsWith("/*") =>
+            sb.append(formatComment(c, state.indentation))
+          case token: Interpolation.Part =>
+            sb.append(formatMarginizedString(token, state.indentation))
+          case literal: Literal.String => // Ignore, see below.
+          case token => sb.append(token.code)
+        }
+        sb.append(whitespace)
+        formatToken.right match {
+          // state.column matches the end of formatToken.right
+          case literal: Literal.String =>
+            val column =
+              if (state.splits.last.modification.isNewline) state.indentation
+              else lastState.column + whitespace.length
+            sb.append(formatMarginizedString(literal, column + 2))
+          case _ => // Ignore
+        }
+        lastState = state
+    }
+    sb.toString()
+  }
+
+  private def formatComment(comment: Comment, indent: Int): String = {
+    val isDocstring = comment.code.startsWith("/**")
+    val spaces: String =
+      if (isDocstring && style.scalaDocs) " " * (indent + 2)
+      else " " * (indent + 1)
+    comment.code.replaceAll("\n *\\*", s"\n$spaces\\*")
+  }
+
+  private def formatMarginizedString(token: Token, indent: Int): String = {
+    if (!style.indentMarginizedStrings) token.code
+    else if (token.isInstanceOf[Interpolation.Part] ||
+        isMarginizedString(token)) {
+      val spaces = " " * indent
+      token.code.replaceAll("\n *\\|", s"\n$spaces\\|")
+    } else {
+      token.code
+    }
+  }
+
+
+}
